refactor(web): move selection/filter logic to TypeScript module
- Extract selection and filter logic from dangerouslySetInnerHTML in render.tsx
- Add packages/web/src/selection.ts with all selection/filter logic
- Import and initialize in packages/web/src/index.ts
- Remove inline script from render.tsx
- No change to UI/UX, improves maintainability and security
